Appends colData in an scMethrix object

Appends colData in an scMethrix object

append_colData(scm = NULL, colData = NULL, name = "Data")

Arguments

scm
scMethrix; Input scMethrix object

colData
dataframe-like or named vector; For a dataframe-like, must contain row names that correspond with the input scMethrix object. For a named vector, vector names must correspond to the row names of the input scMethrix object

name
string; the name of the column for named vector input. Ignored for matrix input

Value

An scMethrix object

Details

Typically used for clustering. Allows additional information to be added to colData in an scMethrix object after the object creation. It does this via a left join on the original colData. Any samples not included in the colData object will be filled with NAs.

Examples

# For dataframe-like input
data('scMethrix_data')
colData <- colData(scMethrix_data)
colData["Cluster"] <- 1:nrow(colData)
scMethrix_data <- append_colData(scMethrix_data,colData)

#> Warning: Colnames of colData already exist in scMethrix object (Sample). These will be overwritten.

colData(scMethrix_data)

#> DataFrame with 4 rows and 2 columns
#>       Sample   Cluster
#>    <integer> <integer>
#> C1         1         1
#> C2         2         2
#> C3         3         3
#> C4         4         4

# For named vector input
data('scMethrix_data')
colData <- c(C1=1,C2=1,C3=1,C4=2)
scMethrix_data <- append_colData(scMethrix_data,colData, name="Cluster")

#> Warning: Colnames of colData already exist in scMethrix object (Cluster). These will be overwritten.

colData(scMethrix_data)

#> DataFrame with 4 rows and 2 columns
#>       Sample   Cluster
#>    <integer> <numeric>
#> C1         1         1
#> C2         2         1
#> C3         3         1
#> C4         4         2