Generates a cluster object for an scMethrix object

Generates a cluster object for an scMethrix object

cluster_scMethrix(
  scm = NULL,
  dist = NULL,
  assay = "score",
  type = c("hierarchical", "partition", "model"),
  colname = "Cluster",
  n_clusters = 1,
  verbose = TRUE,
  ...
)

Arguments

scm
scMethrix; Input scMethrix object. If this is specified the distance matrix will be a generic spearman.dist distance

dist
dist; Optional. A distance matrix generated for an assay. Will use default paramaters for get_distance_matrix.

assay
string; The assay to use. Default is 'score'

type
string; The type of distance metric to use. Available options are 'hierarchical', 'partition', "model". An arbitrary cluster function can be used, and must return a named vector containing integers representing the cluster membership (e.g. c(C1=1,C2=1,C3=1,C4=2)).

colname
string; the name of the colData column that contains the cluster information

n_clusters
integer; the desired number of clusters. This is ignored for model-based clustering

verbose
boolean; flag to output messages or not

...
Additional parameters for the clustering functions

Value

An scMethrix object

Details

Enables multiple methods of clustering to classify samples in an scMethrix object. Either an scMethrix object or a dist object must be provided for clustering.

See also

Examples

data('scMethrix_data')
scMethrix_data <- impute_regions(scMethrix_data)

#> Starting imputation...

#> Imputed in 0.11s

dist <- get_distance_matrix(scMethrix_data,assay = "impute")  

# For a generic clustering function
# The function must return a named vector of integers
fun <- function (dist) {
    fit <- hclust(dist, method="ward.D")
    fit <- cutree(fit, k=2)
    return(fit)
    }

fun(dist) # Example of arbitrary function output

#> C1 C2 C3 C4 
#>  1  1  1  2 

colData(cluster_scMethrix(scMethrix_data, dist = dist, type = fun))

#> DataFrame with 4 rows and 2 columns
#>       Sample   Cluster
#>    <integer> <integer>
#> C1         1         1
#> C2         2         1
#> C3         3         1
#> C4         4         2