Get the pair-wise distance matrix for an assay

Get the pair-wise distance matrix for an assay

get_distance_matrix(
  scm,
  assay = "score",
  type = c("pearson", "spearman", "kendall", "euclidean", "manhattan", "canberra",
    "binary", "minkowski"),
  verbose = TRUE
)

Arguments

scm
scMethrix; Input scMethrix object

assay
string; The assay to use. Default is 'score'

type
string; The type of distance metric to use. Available options are "pearson", "spearman", "tau", "euclidean", "manhattan", "canberra", "binary", "minkowski". An aribitrary distance function can also be used, so long as the input takes just the specified matrix.

verbose
boolean; flag to output messages or not

Value

matrix; the distance matrix

Details

Utilizes mainly the bioDist package to determine various distance metrics to be used for later clustering.

See also

Examples

data('scMethrix_data')
scMethrix_data <- impute_regions(scMethrix_data) 

#> Starting imputation...

#> Imputed in 0.05s

get_distance_matrix(scMethrix_data,assay = "impute") 

#>           C1        C2        C3
#> C2 0.4271783                    
#> C3 0.3147365 0.4083124          
#> C4 0.5233654 0.5556195 0.3753953

# For an arbitrary distance function
library(bioDist)

#> Loading required package: KernSmooth

#> KernSmooth 2.23 loaded
#> Copyright M. P. Wand 1997-2009

fun <- bioDist::spearman.dist
get_distance_matrix(scMethrix_data,assay = "impute",type = bioDist::spearman.dist)

#>           C1        C2        C3
#> C2 0.4713320                    
#> C3 0.3533758 0.3642182          
#> C4 0.6067592 0.5502410 0.3714192