Extract assays from an scMethrix object

Extract assays from an scMethrix object

get_matrix(
  scm = NULL,
  assay = "score",
  add_loci = FALSE,
  in_granges = FALSE,
  order_by_sd = FALSE
)

Arguments

scm
scMethrix; the single cell methylation experiment

assay
string; name of an existing assay. Default = "score"

add_loci
Default FALSE. If TRUE adds CpG position info to the matrix and returns as a data.table

in_granges
Do you want the outcome in GRanges?

order_by_sd
Order output matrix by standard deviation

Value

HDF5Matrix or matrix

Details

Takes scMethrix object and returns the methylation matrix. This will return in the format used by the object (matrix or HDF5matrix).

Examples

data('scMethrix_data')

# Get methylation data
get_matrix(scMethrix_data)

#>         C1  C2  C3  C4
#>   [1,] 100 100 100  NA
#>   [2,]  NA 100  NA 100
#>   [3,]  NA  NA 100  NA
#>   [4,]  NA 100 100 100
#>   [5,]   0  NA   0   0
#>   [6,] 100  NA  NA  NA
#>   [7,]  NA 100 100  NA
#>   [8,] 100  NA 100   0
#>   [9,]  NA 100  NA  NA
#>  [10,] 100 100  NA  NA
#>  [11,]  NA 100  NA  NA
#>  [12,]  NA 100  NA  NA
#>  [13,]   0 100   0  NA
#>  [14,] 100  NA  NA  NA
#>  [15,]  NA 100 100   0
#>  [16,] 100   0 100 100
#>  [17,] 100 100 100  NA
#>  [18,]  NA   0 100 100
#>  [19,] 100 100 100  NA
#>  [20,] 100 100 100 100
#>  [21,] 100 100  NA 100
#>  [22,]  NA  NA 100 100
#>  [23,]  NA  NA  NA 100
#>  [24,] 100 100 100   0
#>  [25,] 100  NA 100 100
#>  [26,] 100  NA 100 100
#>  [27,] 100  NA 100 100
#>  [28,] 100 100 100  NA
#>  [29,]   0  NA   0   0
#>  [30,] 100 100 100  NA
#>  [31,]  NA   0 100  NA
#>  [32,]  NA  NA 100  NA
#>  [33,] 100 100 100  NA
#>  [34,]   0   0   0   0
#>  [35,] 100  NA 100  NA
#>  [36,] 100 100  NA   0
#>  [37,]  NA  NA   0  NA
#>  [38,] 100  NA   0 100
#>  [39,] 100 100 100 100
#>  [40,]   0  NA   0   0
#>  [41,] 100 100 100 100
#>  [42,]  NA  NA 100 100
#>  [43,] 100 100   0   0
#>  [44,] 100  NA 100  NA
#>  [45,]   0  NA  NA  NA
#>  [46,]  NA  NA 100  NA
#>  [47,]   0  NA  NA  NA
#>  [48,] 100  NA 100 100
#>  [49,]   0   0  50   0
#>  [50,] 100  NA  NA 100
#>  [51,]  NA 100  NA 100
#>  [52,]  NA 100   0  50
#>  [53,]   0   0   0 100
#>  [54,] 100   0 100 100
#>  [55,]  NA  NA 100  NA
#>  [56,] 100 100   0 100
#>  [57,] 100   0 100 100
#>  [58,]  NA  NA 100   0
#>  [59,]   0   0   0   0
#>  [60,] 100  NA 100 100
#>  [61,]  NA  NA 100  NA
#>  [62,] 100  NA  NA  NA
#>  [63,]   0  NA  NA 100
#>  [64,]   0  NA   0  NA
#>  [65,] 100  NA  NA  NA
#>  [66,] 100  NA 100 100
#>  [67,]  NA 100 100  NA
#>  [68,] 100 100 100 100
#>  [69,]  NA  NA 100 100
#>  [70,]  NA 100  NA 100
#>  [71,]  NA  NA 100 100
#>  [72,] 100  NA  NA  NA
#>  [73,]   0  NA  NA 100
#>  [74,] 100 100  NA 100
#>  [75,] 100 100 100 100
#>  [76,]  NA  NA 100  NA
#>  [77,]   0  NA  NA  NA
#>  [78,]  NA  NA   0  NA
#>  [79,] 100 100  NA  NA
#>  [80,]  NA   0   0   0
#>  [81,]  NA  NA  NA 100
#>  [82,]  NA  NA 100  NA
#>  [83,]  NA 100  NA  NA
#>  [84,]   0   0   0   0
#>  [85,]  NA   0   0   0
#>  [86,] 100  NA 100 100
#>  [87,] 100 100  NA 100
#>  [88,]   0 100 100   0
#>  [89,] 100  NA  NA 100
#>  [90,] 100   0 100 100
#>  [91,]  NA 100  NA 100
#>  [92,] 100  NA 100  NA
#>  [93,] 100 100  NA 100
#>  [94,]   0  NA   0  NA
#>  [95,]  NA  NA  NA   0
#>  [96,]  NA 100 100 100
#>  [97,] 100  NA 100 100
#>  [98,]  NA  NA 100 100
#>  [99,] 100  NA 100 100
#> [100,]  NA 100 100  NA
#> [101,] 100  NA 100  50
#> [102,] 100  NA 100 100
#> [103,] 100  NA  NA 100
#> [104,]   0   0   0   0
#> [105,]   0   0   0  NA
#> [106,] 100  NA 100 100
#> [107,]  NA  NA 100   0
#> [108,]   0  NA   0   0
#> [109,] 100   0 100 100
#> [110,]  NA  NA  NA 100
#> [111,]  NA  NA  NA   0
#> [112,] 100 100 100 100
#> [113,] 100 100 100 100
#> [114,] 100 100  NA   0
#> [115,] 100  NA  NA  NA
#> [116,] 100 100 100 100
#> [117,]  NA 100 100 100
#> [118,]   0 100 100 100
#> [119,]   0   0  NA  NA
#> [120,]   0   0   0  NA
#> [121,]  NA  NA 100  NA
#> [122,]   0 100 100 100
#> [123,]  NA 100 100 100
#> [124,] 100  NA  NA 100
#> [125,]  NA  NA 100 100
#> [126,]   0 100   0 100
#> [127,]   0   0  NA  NA
#> [128,]   0   0   0 100
#> [129,] 100 100 100 100
#> [130,] 100 100  NA 100
#> [131,]  NA 100   0 100
#> [132,]  NA 100 100 100
#> [133,] 100 100  NA 100
#> [134,]  NA  NA 100 100
#> [135,]   0   0  NA  NA
#> [136,] 100  NA 100 100
#> [137,] 100  NA 100  NA
#> [138,]  NA 100   0  NA
#> [139,] 100 100 100 100
#> [140,]  NA 100 100  NA
#> [141,]  NA  NA  NA 100
#> [142,]  NA 100  NA 100
#> [143,] 100  NA 100 100
#> [144,]  NA  NA   0  NA
#> [145,]  NA 100  NA  NA
#> [146,]  NA   0  NA   0
#> [147,]   0 100  NA  NA
#> [148,] 100  NA  NA  NA
#> [149,] 100  NA  NA  NA
#> [150,]  50   0  NA   0
#> [151,] 100   0  NA   0
#> [152,]  50  NA 100  NA
#> [153,] 100  NA  NA  NA
#> [154,] 100  NA 100 100
#> [155,] 100  NA  NA  NA
#> [156,]  NA 100  NA  NA
#> [157,]   0  NA  NA  NA
#> [158,]  NA  NA 100 100
#> [159,] 100 100  NA 100
#> [160,]  NA 100  50   0
#> [161,] 100 100  NA 100
#> [162,] 100  NA 100  NA
#> [163,] 100 100 100  NA
#> [164,]  NA  NA 100  NA
#> [165,] 100 100 100 100
#> [166,] 100 100 100 100
#> [167,]  NA   0  NA   0
#> [168,]  NA  NA   0  NA
#> [169,]  NA  NA   0  NA
#> [170,]  NA  NA 100  NA
#> [171,]   0   0   0  NA
#> [172,] 100 100 100 100
#> [173,] 100  NA 100 100
#> [174,] 100  NA  NA  NA
#> [175,] 100  NA 100  NA
#> [176,]  NA  NA   0  NA
#> [177,]  NA   0  NA   0
#> [178,] 100   0   0   0
#> [179,]  NA 100  NA  NA
#> [180,] 100 100 100 100
#> [181,]   0  NA  NA  NA
#> [182,]  NA   0  NA  NA
#> [183,]   0  NA   0   0
#> [184,]  NA 100 100 100
#> [185,]  NA 100 100  NA
#> [186,]  NA 100  NA 100
#> [187,] 100  NA  NA  NA
#> [188,]  NA 100  NA   0
#> [189,]  NA  NA  NA 100
#> [190,] 100  50   0   0
#> [191,]  NA 100  NA 100
#> [192,]  NA 100  NA 100
#> [193,]  NA  NA 100  NA
#> [194,]  NA  NA  NA 100
#> [195,]  NA  NA  NA 100
#> [196,]   0   0   0   0
#> [197,]  NA  NA   0  NA
#> [198,] 100  NA   0   0
#> [199,]  NA  NA  NA 100
#> [200,]  NA  NA   0   0
#> [201,]   0   0   0   0
#> [202,]   0   0   0   0
#> [203,]   0   0   0   0
#> [204,]   0   0   0   0
#> [205,]  NA  NA 100  NA
#> [206,]  NA 100  NA  NA
#> [207,]  NA 100 100  NA
#> [208,] 100  NA  NA 100
#> [209,] 100  NA 100  NA
#> [210,] 100 100 100  NA
#> [211,]  NA 100 100 100
#> [212,]   0 100 100   0
#> [213,]  NA 100 100 100
#> [214,] 100 100 100 100
#> [215,]  NA 100  NA  NA
#> [216,]   0   0  NA   0
#> [217,] 100 100  NA  NA
#> [218,] 100  NA  NA   0
#> [219,]  NA 100  NA  NA
#> [220,]  NA  NA 100  50
#> [221,] 100  NA  NA 100
#> [222,]  NA  NA 100  NA
#> [223,]  NA 100  NA  NA
#> [224,]  NA 100   0 100
#> [225,] 100 100 100 100
#> [226,] 100 100  NA 100
#> [227,]   0  NA   0  NA
#> [228,] 100 100 100  NA
#> [229,]   0  NA  NA  NA
#> [230,] 100 100  NA   0
#> [231,] 100 100 100  NA
#> [232,]  NA 100  NA 100
#> [233,] 100   0 100 100
#> [234,] 100 100 100  NA
#> [235,] 100 100 100 100
#> [236,]  NA  NA 100 100
#> [237,] 100  NA  NA  NA
#> [238,] 100 100  NA 100
#> [239,] 100 100   0   0
#> [240,] 100 100   0   0
#> [241,] 100 100 100 100
#> [242,]   0 100  NA   0
#> [243,]  NA 100 100   0
#> [244,] 100  NA 100 100
#> [245,] 100   0   0   0
#> [246,]  NA  NA 100 100
#> [247,] 100   0 100   0
#> [248,]  NA  NA 100 100
#> [249,]  NA  NA  NA 100
#> [250,]   0  NA 100  NA
#> [251,]  NA 100 100  NA
#> [252,] 100 100 100  NA
#> [253,] 100 100  NA  NA
#> [254,] 100  NA  NA  NA
#> [255,] 100  NA 100  NA
#> [256,]   0 100 100  NA
#> [257,] 100 100 100  NA
#> [258,]  NA  NA   0  NA
#> [259,] 100 100 100  NA
#> [260,]  NA 100 100  NA
#> [261,]   0   0 100   0
#> [262,]  NA 100 100 100
#> [263,] 100 100 100 100
#> [264,] 100  NA 100 100
#> [265,] 100  NA 100 100
#> [266,]   0  NA  NA 100
#> [267,] 100 100  NA  NA
#> [268,]  NA  NA 100  NA
#> [269,]  NA  NA  NA   0
#> [270,]   0  NA  NA 100
#> [271,]   0 100  NA 100
#> [272,] 100 100  NA 100
#> [273,] 100 100   0  NA
#> [274,]  50  50 100  NA
#> [275,]  NA  NA   0   0
#> [276,]  NA 100  NA   0
#> [277,] 100 100  NA 100
#> [278,]   0  NA   0   0
#> [279,] 100 100 100 100
#> [280,]  NA  NA  NA   0
#> [281,]   0   0   0   0
#> [282,]   0   0   0   0
#> [283,]   0   0   0   0
#> [284,]   0   0   0   0
#> [285,]   0   0   0   0
#> [286,]   0   0   0   0

# Get methylation data with loci
get_matrix(scMethrix_data, add_loci=TRUE)

#>       chr     start       end  C1  C2  C3  C4
#>   1: chr1   6914480   6914481 100 100 100  NA
#>   2: chr1   9919400   9919401  NA 100  NA 100
#>   3: chr1  13786867  13786868  NA  NA 100  NA
#>   4: chr1  20487719  20487720  NA 100 100 100
#>   5: chr1  20951731  20951732   0  NA   0   0
#>  ---                                         
#> 282: chr2 181520657 181520658   0   0   0   0
#> 283: chr2 181520666 181520667   0   0   0   0
#> 284: chr2 181520679 181520680   0   0   0   0
#> 285: chr2 181520682 181520683   0   0   0   0
#> 286: chr2 181520687 181520688   0   0   0   0

# Get methylation data with loci inside a Granges object 
get_matrix(scMethrix_data, add_loci=TRUE, in_granges=TRUE)

#> GRanges object with 286 ranges and 4 metadata columns:
#>         seqnames              ranges strand |        C1        C2        C3
#>            <Rle>           <IRanges>  <Rle> | <integer> <integer> <integer>
#>     [1]     chr1     6914480-6914481      * |       100       100       100
#>     [2]     chr1     9919400-9919401      * |      <NA>       100      <NA>
#>     [3]     chr1   13786867-13786868      * |      <NA>      <NA>       100
#>     [4]     chr1   20487719-20487720      * |      <NA>       100       100
#>     [5]     chr1   20951731-20951732      * |         0      <NA>         0
#>     ...      ...                 ...    ... .       ...       ...       ...
#>   [282]     chr2 181520657-181520658      * |         0         0         0
#>   [283]     chr2 181520666-181520667      * |         0         0         0
#>   [284]     chr2 181520679-181520680      * |         0         0         0
#>   [285]     chr2 181520682-181520683      * |         0         0         0
#>   [286]     chr2 181520687-181520688      * |         0         0         0
#>                C4
#>         <integer>
#>     [1]      <NA>
#>     [2]       100
#>     [3]      <NA>
#>     [4]       100
#>     [5]         0
#>     ...       ...
#>   [282]         0
#>   [283]         0
#>   [284]         0
#>   [285]         0
#>   [286]         0
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

# Get methylation data sorted by SD
get_matrix(scMethrix_data, order_by_sd = TRUE)

#>         C1  C2  C3  C4
#>   [1,]  NA   0 100  NA
#>   [2,]  NA  NA 100   0
#>   [3,]   0  NA  NA 100
#>   [4,]   0  NA  NA 100
#>   [5,]  NA  NA 100   0
#>   [6,]  NA 100   0  NA
#>   [7,]   0 100  NA  NA
#>   [8,]  NA 100  NA   0
#>   [9,] 100  NA  NA   0
#>  [10,]   0  NA 100  NA
#>  [11,]   0  NA  NA 100
#>  [12,]   0  NA  NA 100
#>  [13,]  NA 100  NA   0
#>  [14,]  NA   0 100 100
#>  [15,] 100  NA   0 100
#>  [16,]  NA 100   0 100
#>  [17,]  NA 100   0 100
#>  [18,]   0 100 100  NA
#>  [19,]   0 100  NA 100
#>  [20,] 100  NA 100   0
#>  [21,]   0 100   0  NA
#>  [22,]  NA 100 100   0
#>  [23,] 100 100  NA   0
#>  [24,] 100 100   0   0
#>  [25,]   0 100 100   0
#>  [26,] 100 100  NA   0
#>  [27,]   0 100   0 100
#>  [28,] 100   0  NA   0
#>  [29,] 100  NA   0   0
#>  [30,]   0 100 100   0
#>  [31,] 100 100  NA   0
#>  [32,] 100 100   0   0
#>  [33,] 100 100   0   0
#>  [34,]   0 100  NA   0
#>  [35,]  NA 100 100   0
#>  [36,] 100   0 100   0
#>  [37,] 100 100   0  NA
#>  [38,] 100   0 100 100
#>  [39,] 100 100 100   0
#>  [40,]  NA 100   0  50
#>  [41,]   0   0   0 100
#>  [42,] 100   0 100 100
#>  [43,] 100 100   0 100
#>  [44,] 100   0 100 100
#>  [45,] 100   0 100 100
#>  [46,] 100   0 100 100
#>  [47,]   0 100 100 100
#>  [48,]   0 100 100 100
#>  [49,]   0   0   0 100
#>  [50,]  NA 100  50   0
#>  [51,] 100   0   0   0
#>  [52,] 100   0 100 100
#>  [53,] 100   0   0   0
#>  [54,]   0   0 100   0
#>  [55,] 100  50   0   0
#>  [56,]  50  NA 100  NA
#>  [57,]  NA  NA 100  50
#>  [58,] 100  NA 100  50
#>  [59,]  50   0  NA   0
#>  [60,]  50  50 100  NA
#>  [61,]   0   0  50   0
#>  [62,] 100 100 100  NA
#>  [63,]  NA 100  NA 100
#>  [64,]  NA 100 100 100
#>  [65,]   0  NA   0   0
#>  [66,]  NA 100 100  NA
#>  [67,] 100 100  NA  NA
#>  [68,] 100 100 100  NA
#>  [69,] 100 100 100  NA
#>  [70,] 100 100 100 100
#>  [71,] 100 100  NA 100
#>  [72,]  NA  NA 100 100
#>  [73,] 100  NA 100 100
#>  [74,] 100  NA 100 100
#>  [75,] 100  NA 100 100
#>  [76,] 100 100 100  NA
#>  [77,]   0  NA   0   0
#>  [78,] 100 100 100  NA
#>  [79,] 100 100 100  NA
#>  [80,]   0   0   0   0
#>  [81,] 100  NA 100  NA
#>  [82,] 100 100 100 100
#>  [83,]   0  NA   0   0
#>  [84,] 100 100 100 100
#>  [85,]  NA  NA 100 100
#>  [86,] 100  NA 100  NA
#>  [87,] 100  NA 100 100
#>  [88,] 100  NA  NA 100
#>  [89,]  NA 100  NA 100
#>  [90,]   0   0   0   0
#>  [91,] 100  NA 100 100
#>  [92,]   0  NA   0  NA
#>  [93,] 100  NA 100 100
#>  [94,]  NA 100 100  NA
#>  [95,] 100 100 100 100
#>  [96,]  NA  NA 100 100
#>  [97,]  NA 100  NA 100
#>  [98,]  NA  NA 100 100
#>  [99,] 100 100  NA 100
#> [100,] 100 100 100 100
#> [101,] 100 100  NA  NA
#> [102,]  NA   0   0   0
#> [103,]   0   0   0   0
#> [104,]  NA   0   0   0
#> [105,] 100  NA 100 100
#> [106,] 100 100  NA 100
#> [107,] 100  NA  NA 100
#> [108,]  NA 100  NA 100
#> [109,] 100  NA 100  NA
#> [110,] 100 100  NA 100
#> [111,]   0  NA   0  NA
#> [112,]  NA 100 100 100
#> [113,] 100  NA 100 100
#> [114,]  NA  NA 100 100
#> [115,] 100  NA 100 100
#> [116,]  NA 100 100  NA
#> [117,] 100  NA 100 100
#> [118,] 100  NA  NA 100
#> [119,]   0   0   0   0
#> [120,]   0   0   0  NA
#> [121,] 100  NA 100 100
#> [122,]   0  NA   0   0
#> [123,] 100 100 100 100
#> [124,] 100 100 100 100
#> [125,] 100 100 100 100
#> [126,]  NA 100 100 100
#> [127,]   0   0  NA  NA
#> [128,]   0   0   0  NA
#> [129,]  NA 100 100 100
#> [130,] 100  NA  NA 100
#> [131,]  NA  NA 100 100
#> [132,]   0   0  NA  NA
#> [133,] 100 100 100 100
#> [134,] 100 100  NA 100
#> [135,]  NA 100 100 100
#> [136,] 100 100  NA 100
#> [137,]  NA  NA 100 100
#> [138,]   0   0  NA  NA
#> [139,] 100  NA 100 100
#> [140,] 100  NA 100  NA
#> [141,] 100 100 100 100
#> [142,]  NA 100 100  NA
#> [143,]  NA 100  NA 100
#> [144,] 100  NA 100 100
#> [145,]  NA   0  NA   0
#> [146,] 100  NA 100 100
#> [147,]  NA  NA 100 100
#> [148,] 100 100  NA 100
#> [149,] 100 100  NA 100
#> [150,] 100  NA 100  NA
#> [151,] 100 100 100  NA
#> [152,] 100 100 100 100
#> [153,] 100 100 100 100
#> [154,]  NA   0  NA   0
#> [155,]   0   0   0  NA
#> [156,] 100 100 100 100
#> [157,] 100  NA 100 100
#> [158,] 100  NA 100  NA
#> [159,]  NA   0  NA   0
#> [160,] 100 100 100 100
#> [161,]   0  NA   0   0
#> [162,]  NA 100 100 100
#> [163,]  NA 100 100  NA
#> [164,]  NA 100  NA 100
#> [165,]  NA 100  NA 100
#> [166,]  NA 100  NA 100
#> [167,]   0   0   0   0
#> [168,]  NA  NA   0   0
#> [169,]   0   0   0   0
#> [170,]   0   0   0   0
#> [171,]   0   0   0   0
#> [172,]   0   0   0   0
#> [173,]  NA 100 100  NA
#> [174,] 100  NA  NA 100
#> [175,] 100  NA 100  NA
#> [176,] 100 100 100  NA
#> [177,]  NA 100 100 100
#> [178,]  NA 100 100 100
#> [179,] 100 100 100 100
#> [180,]   0   0  NA   0
#> [181,] 100 100  NA  NA
#> [182,] 100  NA  NA 100
#> [183,] 100 100 100 100
#> [184,] 100 100  NA 100
#> [185,]   0  NA   0  NA
#> [186,] 100 100 100  NA
#> [187,] 100 100 100  NA
#> [188,]  NA 100  NA 100
#> [189,] 100 100 100  NA
#> [190,] 100 100 100 100
#> [191,]  NA  NA 100 100
#> [192,] 100 100  NA 100
#> [193,] 100 100 100 100
#> [194,] 100  NA 100 100
#> [195,]  NA  NA 100 100
#> [196,]  NA  NA 100 100
#> [197,]  NA 100 100  NA
#> [198,] 100 100 100  NA
#> [199,] 100 100  NA  NA
#> [200,] 100  NA 100  NA
#> [201,] 100 100 100  NA
#> [202,] 100 100 100  NA
#> [203,]  NA 100 100  NA
#> [204,]  NA 100 100 100
#> [205,] 100 100 100 100
#> [206,] 100  NA 100 100
#> [207,] 100  NA 100 100
#> [208,] 100 100  NA  NA
#> [209,] 100 100  NA 100
#> [210,]  NA  NA   0   0
#> [211,] 100 100  NA 100
#> [212,]   0  NA   0   0
#> [213,] 100 100 100 100
#> [214,]   0   0   0   0
#> [215,]   0   0   0   0
#> [216,]   0   0   0   0
#> [217,]   0   0   0   0
#> [218,]   0   0   0   0
#> [219,]   0   0   0   0
#> [220,]  NA  NA 100  NA
#> [221,] 100  NA  NA  NA
#> [222,]  NA 100  NA  NA
#> [223,]  NA 100  NA  NA
#> [224,]  NA 100  NA  NA
#> [225,] 100  NA  NA  NA
#> [226,]  NA  NA  NA 100
#> [227,]  NA  NA 100  NA
#> [228,]  NA  NA   0  NA
#> [229,]   0  NA  NA  NA
#> [230,]  NA  NA 100  NA
#> [231,]   0  NA  NA  NA
#> [232,]  NA  NA 100  NA
#> [233,]  NA  NA 100  NA
#> [234,] 100  NA  NA  NA
#> [235,] 100  NA  NA  NA
#> [236,] 100  NA  NA  NA
#> [237,]  NA  NA 100  NA
#> [238,]   0  NA  NA  NA
#> [239,]  NA  NA   0  NA
#> [240,]  NA  NA  NA 100
#> [241,]  NA  NA 100  NA
#> [242,]  NA 100  NA  NA
#> [243,]  NA  NA  NA   0
#> [244,]  NA  NA  NA 100
#> [245,]  NA  NA  NA   0
#> [246,] 100  NA  NA  NA
#> [247,]  NA  NA 100  NA
#> [248,]  NA  NA  NA 100
#> [249,]  NA  NA   0  NA
#> [250,]  NA 100  NA  NA
#> [251,] 100  NA  NA  NA
#> [252,] 100  NA  NA  NA
#> [253,] 100  NA  NA  NA
#> [254,] 100  NA  NA  NA
#> [255,]  NA 100  NA  NA
#> [256,]   0  NA  NA  NA
#> [257,]  NA  NA 100  NA
#> [258,]  NA  NA   0  NA
#> [259,]  NA  NA   0  NA
#> [260,]  NA  NA 100  NA
#> [261,] 100  NA  NA  NA
#> [262,]  NA  NA   0  NA
#> [263,]  NA 100  NA  NA
#> [264,]   0  NA  NA  NA
#> [265,]  NA   0  NA  NA
#> [266,] 100  NA  NA  NA
#> [267,]  NA  NA  NA 100
#> [268,]  NA  NA 100  NA
#> [269,]  NA  NA  NA 100
#> [270,]  NA  NA  NA 100
#> [271,]  NA  NA   0  NA
#> [272,]  NA  NA  NA 100
#> [273,]  NA  NA 100  NA
#> [274,]  NA 100  NA  NA
#> [275,]  NA 100  NA  NA
#> [276,]  NA 100  NA  NA
#> [277,]  NA  NA 100  NA
#> [278,]  NA 100  NA  NA
#> [279,]   0  NA  NA  NA
#> [280,] 100  NA  NA  NA
#> [281,]  NA  NA  NA 100
#> [282,] 100  NA  NA  NA
#> [283,]  NA  NA   0  NA
#> [284,]  NA  NA 100  NA
#> [285,]  NA  NA  NA   0
#> [286,]  NA  NA  NA   0