Extracts and summarizes methylation or coverage info by regions of interest

Extracts and summarizes methylation or coverage info by regions of interest

get_region_summary(
  scm = NULL,
  regions = NULL,
  n_chunks = 1,
  n_threads = 1,
  type = "score",
  how = "mean",
  overlap_type = "within",
  verbose = TRUE,
  group = NULL
)

Arguments

scm
scMethrix; the single cell methylation experiment

regions
genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a GenomicRanges object

n_chunks
Number of chunks to split the scMethrix object in case it is very large. Default = 1.

n_threads
Number of parallel instances. n_cores should be less than or equal to n_chunks. If n_chunks is not specified, then n_chunks is initialized to be equal to n_cores. Default = 1.

type
string; the assay to be summarized. Default 'score'

how
closure; mathematical function by which regions should be summarized. Can be one of the following: mean, sum, max, min. Default 'mean'

overlap_type
string; defines the type of the overlap of the CpG sites with the target region. Default value is within. For detailed description, see the findOverlaps function of the IRanges package. #param elementMetadata.col columns in scMethrix@elementMetadata which needs to be summarised. Default = NULL.

verbose
boolean; Flag for outputting function status messages. Default = TRUE

group
a column name from sample annotation that defines groups. In this case, the number of min_samples will be tested group-wise.

Value

table of summary statistic for the given region

Details

Takes scMethrix object and summarizes regions

Examples

data('scMethrix_data')
get_region_summary(scMethrix_data,
   regions = data.table(chr = c('chr1','chr2'), start = c(1), end =  c(10000000)),
   type = 'score', how = 'mean')

#> Generating region summary...

#> Summarizing by average

#> Region summary generating in 0.18s

#>     chr start      end n_overlap_CpGs rid        C1  C2  C3  C4
#> 1: chr1     1 10000000              2   1 100.00000 100 100 100
#> 2: chr2     1 10000000              6   2  83.33333   0 100   0