Masks high or low coverage count or cell count

Masks high or low coverage count or cell count

mask_scMethrix(
  scm = NULL,
  low_count = 0,
  high_quantile = NULL,
  n_threads = 1,
  type = "cells"
)

Arguments

scm
scMethrix object

low_count
The minimal coverage allowed. Everything below, will get masked. Default = NULL, nothing gets masked.

high_quantile
The quantile limit of coverage. Quantiles are calculated for each sample and everything that belongs to a higher quantile than the defined will be masked. Default = 0.99.

n_threads
Number of parallel instances. Can only be used if scMethrix is in HDF5 format. Default = 1.

type
Whether to use the "counts" coverage matrix or "cells" count when masking

Value

An object of class scMethrix

Details

Takes scMethrix object and masks sites with too high or too low coverage by putting NA for coverage and beta value. The sites will remain in the object.

Examples

data('scMethrix_data')
mask_scMethrix(scMethrix_data,low_count=4,type="counts")

#> Masking CpG sites in  quintile and counts < 4

#>    Masked 30 CpGs due to too low cells in sample C1.

#>    Masked 25 CpGs due to too low cells in sample C2.

#>    Masked 28 CpGs due to too low cells in sample C3.

#>    Masked 27 CpGs due to too low cells in sample C4.

#> Masked in 0.42s

#> An object of class scMethrix
#>    n_CpGs: 100
#>    n_samples: 4
#>    assays: score, counts
#>    is_h5: FALSE
#>    Reference: hg19
#>    Physical size: 37 Kb