Subsets a given list of CpGs by another list of CpGs

Subsets a given list of CpGs by another list of CpGs

parse_source_idx(
  chr_idx = NULL,
  start_idx = NULL,
  end_idx = NULL,
  strand_idx = NULL,
  beta_idx = NULL,
  M_idx = NULL,
  U_idx = NULL,
  cov_idx = NULL,
  verbose = TRUE
)

Arguments

chr_idx
integer; column of the chromosome

start_idx
integer; column of the CpG start site

end_idx
integer; column of the CpG end site

strand_idx
integer; column of the strand

beta_idx
integer; column of the beta value

M_idx
integer; column of the # of methylated reads

U_idx
integer; column of the # of unmethylated reads

cov_idx
integer; column of the coverage

verbose
flag to output messages or not

Value

List of column names and indexes

Details

Typically used to reduce the number of potential CpG sites to include only those present in the input files so as to maximize performance and minimize resources. Can also be used for quality control to see if there is excessive number of CpG sites that are not present in the reference genome.