Writes values from input BED files into an in-disk HDF5Array

Writes values from input BED files into an in-disk HDF5Array

read_hdf5_data(
  files,
  ref_cpgs,
  n_threads = 0,
  h5_temp = NULL,
  zero_based = FALSE,
  verbose = TRUE,
  meth_idx = 4,
  cov_idx = NULL
)

Arguments

files
The BED files to parse

ref_cpgs
The index of all unique coordinates from the input BED files

n_threads
The number of threads to use. 0 is the default thread with no cluster built.

h5_temp
The file location to store the RealizationSink object

zero_based
Boolean flag for whether the input data is zero-based or not

verbose
flag to output messages or not.

meth_idx
The column index of the methylation value for the read

cov_idx
The column index(es) of the read count

Value

List of HDF5Array. 1 is methylation, 2 is coverage. If no cov_idx is specified, 2 will be NULL

Details

Using the generated index for genomic coordinates, creates a NA-based dense matrtix of methylation values for each BED file/sample. Each column contains the meth. values for a single sample.

Examples

if (FALSE) {
#Do Nothing
}