Extract assays from an scMethrix object

Extract assays from an scMethrix object

get_matrix(scm = NULL, add_loci = FALSE, in_granges = FALSE, type = "score")

Arguments

scm
scMethrix object

add_loci
Default FALSE. If TRUE adds CpG position info to the matrix and returns as a data.table

in_granges
Do you want the outcome in GRanges?

type
Which matrix to get

Value

Coverage or Methylation matrix

Details

Takes scMethrix object and returns the methylation matrix

Examples

data('scMethrix_data')

# Get methylation data
get_matrix(scMethrix_data)

#>         C1  C2  C3  C4
#>   [1,]  NA 100 100 100
#>   [2,]  NA 100 100 100
#>   [3,]  NA  NA  NA 100
#>   [4,] 100  NA  NA  NA
#>   [5,] 100  NA  NA  NA
#>   [6,]  NA 100 100 100
#>   [7,]   0  NA   0   0
#>   [8,]  NA  NA 100  NA
#>   [9,]   0  NA   0   0
#>  [10,] 100 100 100 100
#>  [11,] 100  NA   0 100
#>  [12,] 100 100 100 100
#>  [13,]  NA 100  NA 100
#>  [14,]   0  NA  NA   0
#>  [15,]  NA  NA  NA   0
#>  [16,] 100  NA  NA  NA
#>  [17,] 100 100 100  NA
#>  [18,] 100 100 100  NA
#>  [19,] 100  NA 100   0
#>  [20,]  NA   0 100  NA
#>  [21,]  NA  NA 100 100
#>  [22,] 100 100  NA  NA
#>  [23,] 100 100  NA  NA
#>  [24,]   0   0   0   0
#>  [25,]  NA  NA  NA 100
#>  [26,] 100 100 100 100
#>  [27,] 100  NA 100  NA
#>  [28,]  NA  NA  NA 100
#>  [29,] 100   0 100 100
#>  [30,]  NA 100 100 100
#>  [31,]   0  NA  NA  NA
#>  [32,]   0 100   0  NA
#>  [33,]   0   0   0  15
#>  [34,] 100 100 100  NA
#>  [35,]  NA 100 100   0
#>  [36,] 100 100  NA 100
#>  [37,]  NA 100  NA  NA
#>  [38,]   0  NA   0  NA
#>  [39,]  21  30   6  34
#>  [40,]  14   0   0   0
#>  [41,]  16  20   0   0
#>  [42,]  16  33  33   0
#>  [43,]   0  20  25  33
#>  [44,]  31   3  13   6
#>  [45,] 100 100 100 100
#>  [46,]  NA 100  NA 100
#>  [47,]   0  NA  NA   0
#>  [48,]  NA  NA 100 100
#>  [49,]   0   0 100 100
#>  [50,]  NA  NA  NA 100
#>  [51,]  NA  NA  NA 100
#>  [52,] 100 100 100 100
#>  [53,]  NA  NA   0  NA
#>  [54,]   0  NA   0  NA
#>  [55,]   0  NA  NA  NA
#>  [56,]  NA  NA   0  NA
#>  [57,] 100 100 100 100
#>  [58,]  50 100 100 100
#>  [59,]  NA 100 100  NA
#>  [60,] 100   0   0   0
#>  [61,]   0   0   0   0
#>  [62,] 100  NA 100 100
#>  [63,]  NA  NA  NA 100
#>  [64,] 100 100 100  NA
#>  [65,]  NA  NA 100  NA
#>  [66,]  NA  NA 100 100
#>  [67,]  NA  NA   0   0
#>  [68,] 100 100  NA  NA
#>  [69,]  NA  NA   0   0
#>  [70,] 100  NA 100   0
#>  [71,] 100   0  NA   0
#>  [72,] 100  NA  NA  NA
#>  [73,] 100   0  NA  NA
#>  [74,] 100 100 100 100
#>  [75,] 100  NA 100 100
#>  [76,] 100 100 100 100
#>  [77,]  NA 100  NA  NA
#>  [78,]  NA 100 100  NA
#>  [79,] 100 100 100  NA
#>  [80,]  NA   0 100  NA
#>  [81,]  NA  NA   0   0
#>  [82,]  NA   0   0   0
#>  [83,]  NA 100 100  NA
#>  [84,]   0 100  NA   0
#>  [85,]  NA  NA  NA   0
#>  [86,]   0 100 100  NA
#>  [87,]  NA  NA  NA 100
#>  [88,] 100  NA  NA   0
#>  [89,] 100  NA 100 100
#>  [90,]  NA 100  NA  NA
#>  [91,] 100 100  NA 100
#>  [92,] 100  NA  NA  NA
#>  [93,] 100  NA 100  NA
#>  [94,]  NA 100  NA  NA
#>  [95,] 100 100   0   0
#>  [96,]  50 100  50 100
#>  [97,] 100  NA  NA   0
#>  [98,]  42  33  32   0
#>  [99,] 100 100 100  50
#> [100,]  NA   0  NA  NA

# Get methylation data with loci
get_matrix(scMethrix_data, add_loci=TRUE)

#>                 chr     start       end  C1  C2  C3  C4
#>   1:           chr1   5214622   5214623  NA 100 100 100
#>   2:           chr1 126658854 126658855  NA 100 100 100
#>   3:          chr10  12545268  12545269  NA  NA  NA 100
#>   4:          chr10  12970846  12970847 100  NA  NA  NA
#>   5:          chr10  13182357  13182358 100  NA  NA  NA
#>   6:          chr10  43495896  43495897  NA 100 100 100
#>   7:          chr10  88379338  88379339   0  NA   0   0
#>   8:          chr10  92325422  92325423  NA  NA 100  NA
#>   9:          chr10 126962366 126962367   0  NA   0   0
#>  10:          chr11   8981394   8981395 100 100 100 100
#>  11:          chr11  14876995  14876996 100  NA   0 100
#>  12:          chr11  29254472  29254473 100 100 100 100
#>  13:          chr11  46168321  46168322  NA 100  NA 100
#>  14:          chr11  67820899  67820900   0  NA  NA   0
#>  15:          chr11  68693315  68693316  NA  NA  NA   0
#>  16:          chr11 108897922 108897923 100  NA  NA  NA
#>  17:          chr11 118032702 118032703 100 100 100  NA
#>  18:          chr12  15747383  15747384 100 100 100  NA
#>  19:          chr12  57638220  57638221 100  NA 100   0
#>  20:          chr12  67975902  67975903  NA   0 100  NA
#>  21:          chr12  80742921  80742922  NA  NA 100 100
#>  22:          chr12  91676229  91676230 100 100  NA  NA
#>  23:          chr13  57374206  57374207 100 100  NA  NA
#>  24:          chr13 114458658 114458659   0   0   0   0
#>  25:          chr14   7807618   7807619  NA  NA  NA 100
#>  26:          chr14  19417161  19417162 100 100 100 100
#>  27:          chr14  98616862  98616863 100  NA 100  NA
#>  28:          chr14 120221765 120221766  NA  NA  NA 100
#>  29:          chr15  12477596  12477597 100   0 100 100
#>  30:          chr15  99273706  99273707  NA 100 100 100
#>  31:          chr16  28316291  28316292   0  NA  NA  NA
#>  32:          chr16  38483791  38483792   0 100   0  NA
#>  33:          chr16  57391583  57391584   0   0   0  15
#>  34:          chr16  71406174  71406175 100 100 100  NA
#>  35:          chr16  92138390  92138391  NA 100 100   0
#>  36:          chr17  11548668  11548669 100 100  NA 100
#>  37:          chr17  15112923  15112924  NA 100  NA  NA
#>  38:          chr17  27120761  27120762   0  NA   0  NA
#>  39:          chr17  39843365  39843366  21  30   6  34
#>  40:          chr17  39843484  39843485  14   0   0   0
#>  41:          chr17  39844690  39844691  16  20   0   0
#>  42:          chr17  39845106  39845107  16  33  33   0
#>  43:          chr17  39845925  39845926   0  20  25  33
#>  44:          chr17  39847732  39847733  31   3  13   6
#>  45:          chr17  84815780  84815781 100 100 100 100
#>  46:          chr18  32406094  32406095  NA 100  NA 100
#>  47:          chr18  34861822  34861823   0  NA  NA   0
#>  48:          chr18  35977551  35977552  NA  NA 100 100
#>  49:          chr18  42068023  42068024   0   0 100 100
#>  50:          chr18  44596635  44596636  NA  NA  NA 100
#>  51:          chr18  48679158  48679159  NA  NA  NA 100
#>  52:          chr18  77524693  77524694 100 100 100 100
#>  53:          chr19  33641689  33641690  NA  NA   0  NA
#>  54:          chr19  59469483  59469484   0  NA   0  NA
#>  55:           chr2  21368739  21368740   0  NA  NA  NA
#>  56:           chr2  27515161  27515162  NA  NA   0  NA
#>  57:           chr2  74942217  74942218 100 100 100 100
#>  58:           chr2 102040453 102040454  50 100 100 100
#>  59:           chr2 141098089 141098090  NA 100 100  NA
#>  60:           chr2 168556489 168556490 100   0   0   0
#>  61:           chr2 172345891 172345892   0   0   0   0
#>  62:           chr3  54897928  54897929 100  NA 100 100
#>  63:           chr3  83841279  83841280  NA  NA  NA 100
#>  64:           chr3  95741705  95741706 100 100 100  NA
#>  65:           chr3 108477624 108477625  NA  NA 100  NA
#>  66:           chr3 129242188 129242189  NA  NA 100 100
#>  67:           chr4  20008690  20008691  NA  NA   0   0
#>  68:           chr4  63022673  63022674 100 100  NA  NA
#>  69:           chr4  89688565  89688566  NA  NA   0   0
#>  70:           chr4 128294563 128294564 100  NA 100   0
#>  71:           chr4 135645209 135645210 100   0  NA   0
#>  72:           chr4 137029986 137029987 100  NA  NA  NA
#>  73:           chr4 140557694 140557695 100   0  NA  NA
#>  74:           chr4 143577137 143577138 100 100 100 100
#>  75:           chr4 153638873 153638874 100  NA 100 100
#>  76:           chr5  34563165  34563166 100 100 100 100
#>  77:           chr5  58634484  58634485  NA 100  NA  NA
#>  78:           chr5  75463782  75463783  NA 100 100  NA
#>  79:           chr5 100073190 100073191 100 100 100  NA
#>  80:           chr5 121389180 121389181  NA   0 100  NA
#>  81:           chr5 122156653 122156654  NA  NA   0   0
#>  82:           chr5 127632276 127632277  NA   0   0   0
#>  83:           chr5 148310266 148310267  NA 100 100  NA
#>  84:           chr5 149463664 149463665   0 100  NA   0
#>  85:           chr6  93314028  93314029  NA  NA  NA   0
#>  86:           chr6 116677630 116677631   0 100 100  NA
#>  87:           chr7   4773798   4773799  NA  NA  NA 100
#>  88:           chr7  39767527  39767528 100  NA  NA   0
#>  89:           chr7  46061648  46061649 100  NA 100 100
#>  90:           chr7 114305695 114305696  NA 100  NA  NA
#>  91:           chr7 143042936 143042937 100 100  NA 100
#>  92:           chr8  77415252  77415253 100  NA  NA  NA
#>  93:           chr8 112732216 112732217 100  NA 100  NA
#>  94:           chr8 125602584 125602585  NA 100  NA  NA
#>  95:           chr9  31952251  31952252 100 100   0   0
#>  96:           chr9  35305533  35305534  50 100  50 100
#>  97:           chr9 107951033 107951034 100  NA  NA   0
#>  98:           chr9 123461998 123461999  42  33  32   0
#>  99: chrUn_JH584304     46480     46481 100 100 100  50
#> 100:           chrX  74642207  74642208  NA   0  NA  NA
#>                 chr     start       end  C1  C2  C3  C4

# Get methylation data with loci inside a Granges object 
get_matrix(scMethrix_data, add_loci=TRUE, in_granges=TRUE)

#> GRanges object with 100 ranges and 4 metadata columns:
#>               seqnames              ranges strand |        C1        C2
#>                  <Rle>           <IRanges>  <Rle> | <integer> <integer>
#>     [1]           chr1     5214622-5214623      * |      <NA>       100
#>     [2]           chr1 126658854-126658855      * |      <NA>       100
#>     [3]          chr10   12545268-12545269      * |      <NA>      <NA>
#>     [4]          chr10   12970846-12970847      * |       100      <NA>
#>     [5]          chr10   13182357-13182358      * |       100      <NA>
#>     ...            ...                 ...    ... .       ...       ...
#>    [96]           chr9   35305533-35305534      * |        50       100
#>    [97]           chr9 107951033-107951034      * |       100      <NA>
#>    [98]           chr9 123461998-123461999      * |        42        33
#>    [99] chrUn_JH584304         46480-46481      * |       100       100
#>   [100]           chrX   74642207-74642208      * |      <NA>         0
#>                C3        C4
#>         <integer> <integer>
#>     [1]       100       100
#>     [2]       100       100
#>     [3]      <NA>       100
#>     [4]      <NA>      <NA>
#>     [5]      <NA>      <NA>
#>     ...       ...       ...
#>    [96]        50       100
#>    [97]      <NA>         0
#>    [98]        32         0
#>    [99]       100        50
#>   [100]      <NA>      <NA>
#>   -------
#>   seqinfo: 21 sequences from an unspecified genome; no seqlengths