Extract assays from an scMethrix object

Extract assays from an scMethrix object

get_matrix(
  scm = NULL,
  assay = "score",
  add_loci = FALSE,
  in_granges = FALSE,
  order_by_sd = FALSE,
  n_chunks = 1,
  by = c("row", "column")
)

Arguments

scm
scMethrix; the single cell methylation experiment

assay
string; name of an existing assay. Default = "score"

add_loci
Default FALSE. If TRUE adds CpG position info to the matrix and returns as a data.table

in_granges
Do you want the outcome in GRanges?

order_by_sd
Order output matrix by standard deviation

n_chunks
integer; Number of chunks to split the scMethrix object in case it is very large. Default = 1

by
string; split the matrix by "row" or "col" if n_chunks != 1

Value

HDF5Matrix or matrix

Details

Takes scMethrix object and returns the methylation matrix. This will return in the format used by the object (matrix or HDF5matrix).

Examples

data('scMethrix_data')

# Get methylation data
get_matrix(scMethrix_data)

#>         C1  C2  C3  C4
#>   [1,] 1.0 1.0 1.0  NA
#>   [2,]  NA 1.0  NA 1.0
#>   [3,]  NA  NA 1.0  NA
#>   [4,]  NA 1.0 1.0 1.0
#>   [5,] 0.0  NA 0.0 0.0
#>   [6,] 1.0  NA  NA  NA
#>   [7,]  NA 1.0 1.0  NA
#>   [8,] 1.0  NA 1.0 0.0
#>   [9,]  NA 1.0  NA  NA
#>  [10,] 1.0 1.0  NA  NA
#>  [11,]  NA 1.0  NA  NA
#>  [12,]  NA 1.0  NA  NA
#>  [13,] 0.0 1.0 0.0  NA
#>  [14,] 1.0  NA  NA  NA
#>  [15,]  NA 1.0 1.0 0.0
#>  [16,] 1.0 0.0 1.0 1.0
#>  [17,] 1.0 1.0 1.0  NA
#>  [18,]  NA 0.0 1.0 1.0
#>  [19,] 1.0 1.0 1.0  NA
#>  [20,] 1.0 1.0 1.0 1.0
#>  [21,] 1.0 1.0  NA 1.0
#>  [22,]  NA  NA 1.0 1.0
#>  [23,]  NA  NA  NA 1.0
#>  [24,] 1.0 1.0 1.0 0.0
#>  [25,] 1.0  NA 1.0 1.0
#>  [26,] 1.0  NA 1.0 1.0
#>  [27,] 1.0  NA 1.0 1.0
#>  [28,] 1.0 1.0 1.0  NA
#>  [29,] 0.0  NA 0.0 0.0
#>  [30,] 1.0 1.0 1.0  NA
#>  [31,]  NA 0.0 1.0  NA
#>  [32,]  NA  NA 1.0  NA
#>  [33,] 1.0 1.0 1.0  NA
#>  [34,] 0.0 0.0 0.0 0.0
#>  [35,] 1.0  NA 1.0  NA
#>  [36,] 1.0 1.0  NA 0.0
#>  [37,]  NA  NA 0.0  NA
#>  [38,] 1.0  NA 0.0 1.0
#>  [39,] 1.0 1.0 1.0 1.0
#>  [40,] 0.0  NA 0.0 0.0
#>  [41,] 1.0 1.0 1.0 1.0
#>  [42,]  NA  NA 1.0 1.0
#>  [43,] 1.0 1.0 0.0 0.0
#>  [44,] 1.0  NA 1.0  NA
#>  [45,] 0.0  NA  NA  NA
#>  [46,]  NA  NA 1.0  NA
#>  [47,] 0.0  NA  NA  NA
#>  [48,] 1.0  NA 1.0 1.0
#>  [49,] 0.0 0.0 0.5 0.0
#>  [50,] 1.0  NA  NA 1.0
#>  [51,]  NA 1.0  NA 1.0
#>  [52,]  NA 1.0 0.0 0.5
#>  [53,] 0.0 0.0 0.0 1.0
#>  [54,] 1.0 0.0 1.0 1.0
#>  [55,]  NA  NA 1.0  NA
#>  [56,] 1.0 1.0 0.0 1.0
#>  [57,] 1.0 0.0 1.0 1.0
#>  [58,]  NA  NA 1.0 0.0
#>  [59,] 0.0 0.0 0.0 0.0
#>  [60,] 1.0  NA 1.0 1.0
#>  [61,]  NA  NA 1.0  NA
#>  [62,] 1.0  NA  NA  NA
#>  [63,] 0.0  NA  NA 1.0
#>  [64,] 0.0  NA 0.0  NA
#>  [65,] 1.0  NA  NA  NA
#>  [66,] 1.0  NA 1.0 1.0
#>  [67,]  NA 1.0 1.0  NA
#>  [68,] 1.0 1.0 1.0 1.0
#>  [69,]  NA  NA 1.0 1.0
#>  [70,]  NA 1.0  NA 1.0
#>  [71,]  NA  NA 1.0 1.0
#>  [72,] 1.0  NA  NA  NA
#>  [73,] 0.0  NA  NA 1.0
#>  [74,] 1.0 1.0  NA 1.0
#>  [75,] 1.0 1.0 1.0 1.0
#>  [76,]  NA  NA 1.0  NA
#>  [77,] 0.0  NA  NA  NA
#>  [78,]  NA  NA 0.0  NA
#>  [79,] 1.0 1.0  NA  NA
#>  [80,]  NA 0.0 0.0 0.0
#>  [81,]  NA  NA  NA 1.0
#>  [82,]  NA  NA 1.0  NA
#>  [83,]  NA 1.0  NA  NA
#>  [84,] 0.0 0.0 0.0 0.0
#>  [85,]  NA 0.0 0.0 0.0
#>  [86,] 1.0  NA 1.0 1.0
#>  [87,] 1.0 1.0  NA 1.0
#>  [88,] 0.0 1.0 1.0 0.0
#>  [89,] 1.0  NA  NA 1.0
#>  [90,] 1.0 0.0 1.0 1.0
#>  [91,]  NA 1.0  NA 1.0
#>  [92,] 1.0  NA 1.0  NA
#>  [93,] 1.0 1.0  NA 1.0
#>  [94,] 0.0  NA 0.0  NA
#>  [95,]  NA  NA  NA 0.0
#>  [96,]  NA 1.0 1.0 1.0
#>  [97,] 1.0  NA 1.0 1.0
#>  [98,]  NA  NA 1.0 1.0
#>  [99,] 1.0  NA 1.0 1.0
#> [100,]  NA 1.0 1.0  NA
#> [101,] 1.0  NA 1.0 0.5
#> [102,] 1.0  NA 1.0 1.0
#> [103,] 1.0  NA  NA 1.0
#> [104,] 0.0 0.0 0.0 0.0
#> [105,] 0.0 0.0 0.0  NA
#> [106,] 1.0  NA 1.0 1.0
#> [107,]  NA  NA 1.0 0.0
#> [108,] 0.0  NA 0.0 0.0
#> [109,] 1.0 0.0 1.0 1.0
#> [110,]  NA  NA  NA 1.0
#> [111,]  NA  NA  NA 0.0
#> [112,] 1.0 1.0 1.0 1.0
#> [113,] 1.0 1.0 1.0 1.0
#> [114,] 1.0 1.0  NA 0.0
#> [115,] 1.0  NA  NA  NA
#> [116,] 1.0 1.0 1.0 1.0
#> [117,]  NA 1.0 1.0 1.0
#> [118,] 0.0 1.0 1.0 1.0
#> [119,] 0.0 0.0  NA  NA
#> [120,] 0.0 0.0 0.0  NA
#> [121,]  NA  NA 1.0  NA
#> [122,] 0.0 1.0 1.0 1.0
#> [123,]  NA 1.0 1.0 1.0
#> [124,] 1.0  NA  NA 1.0
#> [125,]  NA  NA 1.0 1.0
#> [126,] 0.0 1.0 0.0 1.0
#> [127,] 0.0 0.0  NA  NA
#> [128,] 0.0 0.0 0.0 1.0
#> [129,] 1.0 1.0 1.0 1.0
#> [130,] 1.0 1.0  NA 1.0
#> [131,]  NA 1.0 0.0 1.0
#> [132,]  NA 1.0 1.0 1.0
#> [133,] 1.0 1.0  NA 1.0
#> [134,]  NA  NA 1.0 1.0
#> [135,] 0.0 0.0  NA  NA
#> [136,] 1.0  NA 1.0 1.0
#> [137,] 1.0  NA 1.0  NA
#> [138,]  NA 1.0 0.0  NA
#> [139,] 1.0 1.0 1.0 1.0
#> [140,]  NA 1.0 1.0  NA
#> [141,]  NA  NA  NA 1.0
#> [142,]  NA 1.0  NA 1.0
#> [143,] 1.0  NA 1.0 1.0
#> [144,]  NA  NA 0.0  NA
#> [145,]  NA 1.0  NA  NA
#> [146,]  NA 0.0  NA 0.0
#> [147,] 0.0 1.0  NA  NA
#> [148,] 1.0  NA  NA  NA
#> [149,] 1.0  NA  NA  NA
#> [150,] 0.5 0.0  NA 0.0
#> [151,] 1.0 0.0  NA 0.0
#> [152,] 0.5  NA 1.0  NA
#> [153,] 1.0  NA  NA  NA
#> [154,] 1.0  NA 1.0 1.0
#> [155,] 1.0  NA  NA  NA
#> [156,]  NA 1.0  NA  NA
#> [157,] 0.0  NA  NA  NA
#> [158,]  NA  NA 1.0 1.0
#> [159,] 1.0 1.0  NA 1.0
#> [160,]  NA 1.0 0.5 0.0
#> [161,] 1.0 1.0  NA 1.0
#> [162,] 1.0  NA 1.0  NA
#> [163,] 1.0 1.0 1.0  NA
#> [164,]  NA  NA 1.0  NA
#> [165,] 1.0 1.0 1.0 1.0
#> [166,] 1.0 1.0 1.0 1.0
#> [167,]  NA 0.0  NA 0.0
#> [168,]  NA  NA 0.0  NA
#> [169,]  NA  NA 0.0  NA
#> [170,]  NA  NA 1.0  NA
#> [171,] 0.0 0.0 0.0  NA
#> [172,] 1.0 1.0 1.0 1.0
#> [173,] 1.0  NA 1.0 1.0
#> [174,] 1.0  NA  NA  NA
#> [175,] 1.0  NA 1.0  NA
#> [176,]  NA  NA 0.0  NA
#> [177,]  NA 0.0  NA 0.0
#> [178,] 1.0 0.0 0.0 0.0
#> [179,]  NA 1.0  NA  NA
#> [180,] 1.0 1.0 1.0 1.0
#> [181,] 0.0  NA  NA  NA
#> [182,]  NA 0.0  NA  NA
#> [183,] 0.0  NA 0.0 0.0
#> [184,]  NA 1.0 1.0 1.0
#> [185,]  NA 1.0 1.0  NA
#> [186,]  NA 1.0  NA 1.0
#> [187,] 1.0  NA  NA  NA
#> [188,]  NA 1.0  NA 0.0
#> [189,]  NA  NA  NA 1.0
#> [190,] 1.0 0.5 0.0 0.0
#> [191,]  NA 1.0  NA 1.0
#> [192,]  NA 1.0  NA 1.0
#> [193,]  NA  NA 1.0  NA
#> [194,]  NA  NA  NA 1.0
#> [195,]  NA  NA  NA 1.0
#> [196,] 0.0 0.0 0.0 0.0
#> [197,]  NA  NA 0.0  NA
#> [198,] 1.0  NA 0.0 0.0
#> [199,]  NA  NA  NA 1.0
#> [200,]  NA  NA 0.0 0.0
#> [201,] 0.0 0.0 0.0 0.0
#> [202,] 0.0 0.0 0.0 0.0
#> [203,] 0.0 0.0 0.0 0.0
#> [204,] 0.0 0.0 0.0 0.0
#> [205,]  NA  NA 1.0  NA
#> [206,]  NA 1.0  NA  NA
#> [207,]  NA 1.0 1.0  NA
#> [208,] 1.0  NA  NA 1.0
#> [209,] 1.0  NA 1.0  NA
#> [210,] 1.0 1.0 1.0  NA
#> [211,]  NA 1.0 1.0 1.0
#> [212,] 0.0 1.0 1.0 0.0
#> [213,]  NA 1.0 1.0 1.0
#> [214,] 1.0 1.0 1.0 1.0
#> [215,]  NA 1.0  NA  NA
#> [216,] 0.0 0.0  NA 0.0
#> [217,] 1.0 1.0  NA  NA
#> [218,] 1.0  NA  NA 0.0
#> [219,]  NA 1.0  NA  NA
#> [220,]  NA  NA 1.0 0.5
#> [221,] 1.0  NA  NA 1.0
#> [222,]  NA  NA 1.0  NA
#> [223,]  NA 1.0  NA  NA
#> [224,]  NA 1.0 0.0 1.0
#> [225,] 1.0 1.0 1.0 1.0
#> [226,] 1.0 1.0  NA 1.0
#> [227,] 0.0  NA 0.0  NA
#> [228,] 1.0 1.0 1.0  NA
#> [229,] 0.0  NA  NA  NA
#> [230,] 1.0 1.0  NA 0.0
#> [231,] 1.0 1.0 1.0  NA
#> [232,]  NA 1.0  NA 1.0
#> [233,] 1.0 0.0 1.0 1.0
#> [234,] 1.0 1.0 1.0  NA
#> [235,] 1.0 1.0 1.0 1.0
#> [236,]  NA  NA 1.0 1.0
#> [237,] 1.0  NA  NA  NA
#> [238,] 1.0 1.0  NA 1.0
#> [239,] 1.0 1.0 0.0 0.0
#> [240,] 1.0 1.0 0.0 0.0
#> [241,] 1.0 1.0 1.0 1.0
#> [242,] 0.0 1.0  NA 0.0
#> [243,]  NA 1.0 1.0 0.0
#> [244,] 1.0  NA 1.0 1.0
#> [245,] 1.0 0.0 0.0 0.0
#> [246,]  NA  NA 1.0 1.0
#> [247,] 1.0 0.0 1.0 0.0
#> [248,]  NA  NA 1.0 1.0
#> [249,]  NA  NA  NA 1.0
#> [250,] 0.0  NA 1.0  NA
#> [251,]  NA 1.0 1.0  NA
#> [252,] 1.0 1.0 1.0  NA
#> [253,] 1.0 1.0  NA  NA
#> [254,] 1.0  NA  NA  NA
#> [255,] 1.0  NA 1.0  NA
#> [256,] 0.0 1.0 1.0  NA
#> [257,] 1.0 1.0 1.0  NA
#> [258,]  NA  NA 0.0  NA
#> [259,] 1.0 1.0 1.0  NA
#> [260,]  NA 1.0 1.0  NA
#> [261,] 0.0 0.0 1.0 0.0
#> [262,]  NA 1.0 1.0 1.0
#> [263,] 1.0 1.0 1.0 1.0
#> [264,] 1.0  NA 1.0 1.0
#> [265,] 1.0  NA 1.0 1.0
#> [266,] 0.0  NA  NA 1.0
#> [267,] 1.0 1.0  NA  NA
#> [268,]  NA  NA 1.0  NA
#> [269,]  NA  NA  NA 0.0
#> [270,] 0.0  NA  NA 1.0
#> [271,] 0.0 1.0  NA 1.0
#> [272,] 1.0 1.0  NA 1.0
#> [273,] 1.0 1.0 0.0  NA
#> [274,] 0.5 0.5 1.0  NA
#> [275,]  NA  NA 0.0 0.0
#> [276,]  NA 1.0  NA 0.0
#> [277,] 1.0 1.0  NA 1.0
#> [278,] 0.0  NA 0.0 0.0
#> [279,] 1.0 1.0 1.0 1.0
#> [280,]  NA  NA  NA 0.0
#> [281,] 0.0 0.0 0.0 0.0
#> [282,] 0.0 0.0 0.0 0.0
#> [283,] 0.0 0.0 0.0 0.0
#> [284,] 0.0 0.0 0.0 0.0
#> [285,] 0.0 0.0 0.0 0.0
#> [286,] 0.0 0.0 0.0 0.0

# Get methylation data with loci
get_matrix(scMethrix_data, add_loci=TRUE)

#>       chr     start       end C1 C2 C3 C4
#>   1: chr1   6914480   6914481  1  1  1 NA
#>   2: chr1   9919400   9919401 NA  1 NA  1
#>   3: chr1  13786867  13786868 NA NA  1 NA
#>   4: chr1  20487719  20487720 NA  1  1  1
#>   5: chr1  20951731  20951732  0 NA  0  0
#>  ---                                     
#> 282: chr2 181520657 181520658  0  0  0  0
#> 283: chr2 181520666 181520667  0  0  0  0
#> 284: chr2 181520679 181520680  0  0  0  0
#> 285: chr2 181520682 181520683  0  0  0  0
#> 286: chr2 181520687 181520688  0  0  0  0

# Get methylation data with loci inside a Granges object 
get_matrix(scMethrix_data, add_loci=TRUE, in_granges=TRUE)

#> GRanges object with 286 ranges and 4 metadata columns:
#>         seqnames              ranges strand |        C1        C2        C3
#>            <Rle>           <IRanges>  <Rle> | <numeric> <numeric> <numeric>
#>     [1]     chr1     6914480-6914481      * |         1         1         1
#>     [2]     chr1     9919400-9919401      * |        NA         1        NA
#>     [3]     chr1   13786867-13786868      * |        NA        NA         1
#>     [4]     chr1   20487719-20487720      * |        NA         1         1
#>     [5]     chr1   20951731-20951732      * |         0        NA         0
#>     ...      ...                 ...    ... .       ...       ...       ...
#>   [282]     chr2 181520657-181520658      * |         0         0         0
#>   [283]     chr2 181520666-181520667      * |         0         0         0
#>   [284]     chr2 181520679-181520680      * |         0         0         0
#>   [285]     chr2 181520682-181520683      * |         0         0         0
#>   [286]     chr2 181520687-181520688      * |         0         0         0
#>                C4
#>         <numeric>
#>     [1]        NA
#>     [2]         1
#>     [3]        NA
#>     [4]         1
#>     [5]         0
#>     ...       ...
#>   [282]         0
#>   [283]         0
#>   [284]         0
#>   [285]         0
#>   [286]         0
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

# Get methylation data sorted by SD
get_matrix(scMethrix_data, order_by_sd = TRUE)

#>         C1  C2  C3  C4
#>   [1,]  NA 0.0 1.0  NA
#>   [2,]  NA  NA 1.0 0.0
#>   [3,] 0.0  NA  NA 1.0
#>   [4,] 0.0  NA  NA 1.0
#>   [5,]  NA  NA 1.0 0.0
#>   [6,]  NA 1.0 0.0  NA
#>   [7,] 0.0 1.0  NA  NA
#>   [8,]  NA 1.0  NA 0.0
#>   [9,] 1.0  NA  NA 0.0
#>  [10,] 0.0  NA 1.0  NA
#>  [11,] 0.0  NA  NA 1.0
#>  [12,] 0.0  NA  NA 1.0
#>  [13,]  NA 1.0  NA 0.0
#>  [14,] 1.0  NA 1.0 0.0
#>  [15,] 0.0 1.0 0.0  NA
#>  [16,]  NA 1.0 1.0 0.0
#>  [17,]  NA 0.0 1.0 1.0
#>  [18,] 1.0 1.0  NA 0.0
#>  [19,] 1.0  NA 0.0 1.0
#>  [20,] 1.0 1.0  NA 0.0
#>  [21,]  NA 1.0 0.0 1.0
#>  [22,] 1.0 0.0  NA 0.0
#>  [23,] 1.0  NA 0.0 0.0
#>  [24,]  NA 1.0 0.0 1.0
#>  [25,] 1.0 1.0  NA 0.0
#>  [26,] 0.0 1.0  NA 0.0
#>  [27,]  NA 1.0 1.0 0.0
#>  [28,] 0.0 1.0 1.0  NA
#>  [29,] 0.0 1.0  NA 1.0
#>  [30,] 1.0 1.0 0.0  NA
#>  [31,] 1.0 1.0 0.0 0.0
#>  [32,] 0.0 1.0 1.0 0.0
#>  [33,] 0.0 1.0 0.0 1.0
#>  [34,] 0.0 1.0 1.0 0.0
#>  [35,] 1.0 1.0 0.0 0.0
#>  [36,] 1.0 1.0 0.0 0.0
#>  [37,] 1.0 0.0 1.0 0.0
#>  [38,] 1.0 0.0 1.0 1.0
#>  [39,] 1.0 1.0 1.0 0.0
#>  [40,]  NA 1.0 0.0 0.5
#>  [41,] 0.0 0.0 0.0 1.0
#>  [42,] 1.0 0.0 1.0 1.0
#>  [43,] 1.0 1.0 0.0 1.0
#>  [44,] 1.0 0.0 1.0 1.0
#>  [45,] 1.0 0.0 1.0 1.0
#>  [46,] 1.0 0.0 1.0 1.0
#>  [47,] 0.0 1.0 1.0 1.0
#>  [48,] 0.0 1.0 1.0 1.0
#>  [49,] 0.0 0.0 0.0 1.0
#>  [50,]  NA 1.0 0.5 0.0
#>  [51,] 1.0 0.0 0.0 0.0
#>  [52,] 1.0 0.0 1.0 1.0
#>  [53,] 1.0 0.0 0.0 0.0
#>  [54,] 0.0 0.0 1.0 0.0
#>  [55,] 1.0 0.5 0.0 0.0
#>  [56,] 0.5  NA 1.0  NA
#>  [57,]  NA  NA 1.0 0.5
#>  [58,] 0.5 0.0  NA 0.0
#>  [59,] 1.0  NA 1.0 0.5
#>  [60,] 0.5 0.5 1.0  NA
#>  [61,] 0.0 0.0 0.5 0.0
#>  [62,] 1.0 1.0 1.0  NA
#>  [63,]  NA 1.0  NA 1.0
#>  [64,]  NA 1.0 1.0 1.0
#>  [65,] 0.0  NA 0.0 0.0
#>  [66,]  NA 1.0 1.0  NA
#>  [67,] 1.0 1.0  NA  NA
#>  [68,] 1.0 1.0 1.0  NA
#>  [69,] 1.0 1.0 1.0  NA
#>  [70,] 1.0 1.0 1.0 1.0
#>  [71,] 1.0 1.0  NA 1.0
#>  [72,]  NA  NA 1.0 1.0
#>  [73,] 1.0  NA 1.0 1.0
#>  [74,] 1.0  NA 1.0 1.0
#>  [75,] 1.0  NA 1.0 1.0
#>  [76,] 1.0 1.0 1.0  NA
#>  [77,] 0.0  NA 0.0 0.0
#>  [78,] 1.0 1.0 1.0  NA
#>  [79,] 1.0 1.0 1.0  NA
#>  [80,] 0.0 0.0 0.0 0.0
#>  [81,] 1.0  NA 1.0  NA
#>  [82,] 1.0 1.0 1.0 1.0
#>  [83,] 0.0  NA 0.0 0.0
#>  [84,] 1.0 1.0 1.0 1.0
#>  [85,]  NA  NA 1.0 1.0
#>  [86,] 1.0  NA 1.0  NA
#>  [87,] 1.0  NA 1.0 1.0
#>  [88,] 1.0  NA  NA 1.0
#>  [89,]  NA 1.0  NA 1.0
#>  [90,] 0.0 0.0 0.0 0.0
#>  [91,] 1.0  NA 1.0 1.0
#>  [92,] 0.0  NA 0.0  NA
#>  [93,] 1.0  NA 1.0 1.0
#>  [94,]  NA 1.0 1.0  NA
#>  [95,] 1.0 1.0 1.0 1.0
#>  [96,]  NA  NA 1.0 1.0
#>  [97,]  NA 1.0  NA 1.0
#>  [98,]  NA  NA 1.0 1.0
#>  [99,] 1.0 1.0  NA 1.0
#> [100,] 1.0 1.0 1.0 1.0
#> [101,] 1.0 1.0  NA  NA
#> [102,]  NA 0.0 0.0 0.0
#> [103,] 0.0 0.0 0.0 0.0
#> [104,]  NA 0.0 0.0 0.0
#> [105,] 1.0  NA 1.0 1.0
#> [106,] 1.0 1.0  NA 1.0
#> [107,] 1.0  NA  NA 1.0
#> [108,]  NA 1.0  NA 1.0
#> [109,] 1.0  NA 1.0  NA
#> [110,] 1.0 1.0  NA 1.0
#> [111,] 0.0  NA 0.0  NA
#> [112,]  NA 1.0 1.0 1.0
#> [113,] 1.0  NA 1.0 1.0
#> [114,]  NA  NA 1.0 1.0
#> [115,] 1.0  NA 1.0 1.0
#> [116,]  NA 1.0 1.0  NA
#> [117,] 1.0  NA 1.0 1.0
#> [118,] 1.0  NA  NA 1.0
#> [119,] 0.0 0.0 0.0 0.0
#> [120,] 0.0 0.0 0.0  NA
#> [121,] 1.0  NA 1.0 1.0
#> [122,] 0.0  NA 0.0 0.0
#> [123,] 1.0 1.0 1.0 1.0
#> [124,] 1.0 1.0 1.0 1.0
#> [125,] 1.0 1.0 1.0 1.0
#> [126,]  NA 1.0 1.0 1.0
#> [127,] 0.0 0.0  NA  NA
#> [128,] 0.0 0.0 0.0  NA
#> [129,]  NA 1.0 1.0 1.0
#> [130,] 1.0  NA  NA 1.0
#> [131,]  NA  NA 1.0 1.0
#> [132,] 0.0 0.0  NA  NA
#> [133,] 1.0 1.0 1.0 1.0
#> [134,] 1.0 1.0  NA 1.0
#> [135,]  NA 1.0 1.0 1.0
#> [136,] 1.0 1.0  NA 1.0
#> [137,]  NA  NA 1.0 1.0
#> [138,] 0.0 0.0  NA  NA
#> [139,] 1.0  NA 1.0 1.0
#> [140,] 1.0  NA 1.0  NA
#> [141,] 1.0 1.0 1.0 1.0
#> [142,]  NA 1.0 1.0  NA
#> [143,]  NA 1.0  NA 1.0
#> [144,] 1.0  NA 1.0 1.0
#> [145,]  NA 0.0  NA 0.0
#> [146,] 1.0  NA 1.0 1.0
#> [147,]  NA  NA 1.0 1.0
#> [148,] 1.0 1.0  NA 1.0
#> [149,] 1.0 1.0  NA 1.0
#> [150,] 1.0  NA 1.0  NA
#> [151,] 1.0 1.0 1.0  NA
#> [152,] 1.0 1.0 1.0 1.0
#> [153,] 1.0 1.0 1.0 1.0
#> [154,]  NA 0.0  NA 0.0
#> [155,] 0.0 0.0 0.0  NA
#> [156,] 1.0 1.0 1.0 1.0
#> [157,] 1.0  NA 1.0 1.0
#> [158,] 1.0  NA 1.0  NA
#> [159,]  NA 0.0  NA 0.0
#> [160,] 1.0 1.0 1.0 1.0
#> [161,] 0.0  NA 0.0 0.0
#> [162,]  NA 1.0 1.0 1.0
#> [163,]  NA 1.0 1.0  NA
#> [164,]  NA 1.0  NA 1.0
#> [165,]  NA 1.0  NA 1.0
#> [166,]  NA 1.0  NA 1.0
#> [167,] 0.0 0.0 0.0 0.0
#> [168,]  NA  NA 0.0 0.0
#> [169,] 0.0 0.0 0.0 0.0
#> [170,] 0.0 0.0 0.0 0.0
#> [171,] 0.0 0.0 0.0 0.0
#> [172,] 0.0 0.0 0.0 0.0
#> [173,]  NA 1.0 1.0  NA
#> [174,] 1.0  NA  NA 1.0
#> [175,] 1.0  NA 1.0  NA
#> [176,] 1.0 1.0 1.0  NA
#> [177,]  NA 1.0 1.0 1.0
#> [178,]  NA 1.0 1.0 1.0
#> [179,] 1.0 1.0 1.0 1.0
#> [180,] 0.0 0.0  NA 0.0
#> [181,] 1.0 1.0  NA  NA
#> [182,] 1.0  NA  NA 1.0
#> [183,] 1.0 1.0 1.0 1.0
#> [184,] 1.0 1.0  NA 1.0
#> [185,] 0.0  NA 0.0  NA
#> [186,] 1.0 1.0 1.0  NA
#> [187,] 1.0 1.0 1.0  NA
#> [188,]  NA 1.0  NA 1.0
#> [189,] 1.0 1.0 1.0  NA
#> [190,] 1.0 1.0 1.0 1.0
#> [191,]  NA  NA 1.0 1.0
#> [192,] 1.0 1.0  NA 1.0
#> [193,] 1.0 1.0 1.0 1.0
#> [194,] 1.0  NA 1.0 1.0
#> [195,]  NA  NA 1.0 1.0
#> [196,]  NA  NA 1.0 1.0
#> [197,]  NA 1.0 1.0  NA
#> [198,] 1.0 1.0 1.0  NA
#> [199,] 1.0 1.0  NA  NA
#> [200,] 1.0  NA 1.0  NA
#> [201,] 1.0 1.0 1.0  NA
#> [202,] 1.0 1.0 1.0  NA
#> [203,]  NA 1.0 1.0  NA
#> [204,]  NA 1.0 1.0 1.0
#> [205,] 1.0 1.0 1.0 1.0
#> [206,] 1.0  NA 1.0 1.0
#> [207,] 1.0  NA 1.0 1.0
#> [208,] 1.0 1.0  NA  NA
#> [209,] 1.0 1.0  NA 1.0
#> [210,]  NA  NA 0.0 0.0
#> [211,] 1.0 1.0  NA 1.0
#> [212,] 0.0  NA 0.0 0.0
#> [213,] 1.0 1.0 1.0 1.0
#> [214,] 0.0 0.0 0.0 0.0
#> [215,] 0.0 0.0 0.0 0.0
#> [216,] 0.0 0.0 0.0 0.0
#> [217,] 0.0 0.0 0.0 0.0
#> [218,] 0.0 0.0 0.0 0.0
#> [219,] 0.0 0.0 0.0 0.0
#> [220,]  NA  NA 1.0  NA
#> [221,] 1.0  NA  NA  NA
#> [222,]  NA 1.0  NA  NA
#> [223,]  NA 1.0  NA  NA
#> [224,]  NA 1.0  NA  NA
#> [225,] 1.0  NA  NA  NA
#> [226,]  NA  NA  NA 1.0
#> [227,]  NA  NA 1.0  NA
#> [228,]  NA  NA 0.0  NA
#> [229,] 0.0  NA  NA  NA
#> [230,]  NA  NA 1.0  NA
#> [231,] 0.0  NA  NA  NA
#> [232,]  NA  NA 1.0  NA
#> [233,]  NA  NA 1.0  NA
#> [234,] 1.0  NA  NA  NA
#> [235,] 1.0  NA  NA  NA
#> [236,] 1.0  NA  NA  NA
#> [237,]  NA  NA 1.0  NA
#> [238,] 0.0  NA  NA  NA
#> [239,]  NA  NA 0.0  NA
#> [240,]  NA  NA  NA 1.0
#> [241,]  NA  NA 1.0  NA
#> [242,]  NA 1.0  NA  NA
#> [243,]  NA  NA  NA 0.0
#> [244,]  NA  NA  NA 1.0
#> [245,]  NA  NA  NA 0.0
#> [246,] 1.0  NA  NA  NA
#> [247,]  NA  NA 1.0  NA
#> [248,]  NA  NA  NA 1.0
#> [249,]  NA  NA 0.0  NA
#> [250,]  NA 1.0  NA  NA
#> [251,] 1.0  NA  NA  NA
#> [252,] 1.0  NA  NA  NA
#> [253,] 1.0  NA  NA  NA
#> [254,] 1.0  NA  NA  NA
#> [255,]  NA 1.0  NA  NA
#> [256,] 0.0  NA  NA  NA
#> [257,]  NA  NA 1.0  NA
#> [258,]  NA  NA 0.0  NA
#> [259,]  NA  NA 0.0  NA
#> [260,]  NA  NA 1.0  NA
#> [261,] 1.0  NA  NA  NA
#> [262,]  NA  NA 0.0  NA
#> [263,]  NA 1.0  NA  NA
#> [264,] 0.0  NA  NA  NA
#> [265,]  NA 0.0  NA  NA
#> [266,] 1.0  NA  NA  NA
#> [267,]  NA  NA  NA 1.0
#> [268,]  NA  NA 1.0  NA
#> [269,]  NA  NA  NA 1.0
#> [270,]  NA  NA  NA 1.0
#> [271,]  NA  NA 0.0  NA
#> [272,]  NA  NA  NA 1.0
#> [273,]  NA  NA 1.0  NA
#> [274,]  NA 1.0  NA  NA
#> [275,]  NA 1.0  NA  NA
#> [276,]  NA 1.0  NA  NA
#> [277,]  NA  NA 1.0  NA
#> [278,]  NA 1.0  NA  NA
#> [279,] 0.0  NA  NA  NA
#> [280,] 1.0  NA  NA  NA
#> [281,]  NA  NA  NA 1.0
#> [282,] 1.0  NA  NA  NA
#> [283,]  NA  NA 0.0  NA
#> [284,]  NA  NA 1.0  NA
#> [285,]  NA  NA  NA 0.0
#> [286,]  NA  NA  NA 0.0

# Split the matrix into parts
get_matrix(scMethrix_data, n_chunks = 4, by="row")

#> [[1]]
#>       C1 C2  C3  C4
#>  [1,]  1  1 1.0  NA
#>  [2,] NA  1  NA 1.0
#>  [3,] NA NA 1.0  NA
#>  [4,] NA  1 1.0 1.0
#>  [5,]  0 NA 0.0 0.0
#>  [6,]  1 NA  NA  NA
#>  [7,] NA  1 1.0  NA
#>  [8,]  1 NA 1.0 0.0
#>  [9,] NA  1  NA  NA
#> [10,]  1  1  NA  NA
#> [11,] NA  1  NA  NA
#> [12,] NA  1  NA  NA
#> [13,]  0  1 0.0  NA
#> [14,]  1 NA  NA  NA
#> [15,] NA  1 1.0 0.0
#> [16,]  1  0 1.0 1.0
#> [17,]  1  1 1.0  NA
#> [18,] NA  0 1.0 1.0
#> [19,]  1  1 1.0  NA
#> [20,]  1  1 1.0 1.0
#> [21,]  1  1  NA 1.0
#> [22,] NA NA 1.0 1.0
#> [23,] NA NA  NA 1.0
#> [24,]  1  1 1.0 0.0
#> [25,]  1 NA 1.0 1.0
#> [26,]  1 NA 1.0 1.0
#> [27,]  1 NA 1.0 1.0
#> [28,]  1  1 1.0  NA
#> [29,]  0 NA 0.0 0.0
#> [30,]  1  1 1.0  NA
#> [31,] NA  0 1.0  NA
#> [32,] NA NA 1.0  NA
#> [33,]  1  1 1.0  NA
#> [34,]  0  0 0.0 0.0
#> [35,]  1 NA 1.0  NA
#> [36,]  1  1  NA 0.0
#> [37,] NA NA 0.0  NA
#> [38,]  1 NA 0.0 1.0
#> [39,]  1  1 1.0 1.0
#> [40,]  0 NA 0.0 0.0
#> [41,]  1  1 1.0 1.0
#> [42,] NA NA 1.0 1.0
#> [43,]  1  1 0.0 0.0
#> [44,]  1 NA 1.0  NA
#> [45,]  0 NA  NA  NA
#> [46,] NA NA 1.0  NA
#> [47,]  0 NA  NA  NA
#> [48,]  1 NA 1.0 1.0
#> [49,]  0  0 0.5 0.0
#> [50,]  1 NA  NA 1.0
#> [51,] NA  1  NA 1.0
#> [52,] NA  1 0.0 0.5
#> [53,]  0  0 0.0 1.0
#> [54,]  1  0 1.0 1.0
#> [55,] NA NA 1.0  NA
#> [56,]  1  1 0.0 1.0
#> [57,]  1  0 1.0 1.0
#> [58,] NA NA 1.0 0.0
#> [59,]  0  0 0.0 0.0
#> [60,]  1 NA 1.0 1.0
#> [61,] NA NA 1.0  NA
#> [62,]  1 NA  NA  NA
#> [63,]  0 NA  NA 1.0
#> [64,]  0 NA 0.0  NA
#> [65,]  1 NA  NA  NA
#> [66,]  1 NA 1.0 1.0
#> [67,] NA  1 1.0  NA
#> [68,]  1  1 1.0 1.0
#> [69,] NA NA 1.0 1.0
#> [70,] NA  1  NA 1.0
#> [71,] NA NA 1.0 1.0
#> [72,]  1 NA  NA  NA
#> 
#> [[2]]
#>       C1 C2 C3  C4
#>  [1,]  0 NA NA 1.0
#>  [2,]  1  1 NA 1.0
#>  [3,]  1  1  1 1.0
#>  [4,] NA NA  1  NA
#>  [5,]  0 NA NA  NA
#>  [6,] NA NA  0  NA
#>  [7,]  1  1 NA  NA
#>  [8,] NA  0  0 0.0
#>  [9,] NA NA NA 1.0
#> [10,] NA NA  1  NA
#> [11,] NA  1 NA  NA
#> [12,]  0  0  0 0.0
#> [13,] NA  0  0 0.0
#> [14,]  1 NA  1 1.0
#> [15,]  1  1 NA 1.0
#> [16,]  0  1  1 0.0
#> [17,]  1 NA NA 1.0
#> [18,]  1  0  1 1.0
#> [19,] NA  1 NA 1.0
#> [20,]  1 NA  1  NA
#> [21,]  1  1 NA 1.0
#> [22,]  0 NA  0  NA
#> [23,] NA NA NA 0.0
#> [24,] NA  1  1 1.0
#> [25,]  1 NA  1 1.0
#> [26,] NA NA  1 1.0
#> [27,]  1 NA  1 1.0
#> [28,] NA  1  1  NA
#> [29,]  1 NA  1 0.5
#> [30,]  1 NA  1 1.0
#> [31,]  1 NA NA 1.0
#> [32,]  0  0  0 0.0
#> [33,]  0  0  0  NA
#> [34,]  1 NA  1 1.0
#> [35,] NA NA  1 0.0
#> [36,]  0 NA  0 0.0
#> [37,]  1  0  1 1.0
#> [38,] NA NA NA 1.0
#> [39,] NA NA NA 0.0
#> [40,]  1  1  1 1.0
#> [41,]  1  1  1 1.0
#> [42,]  1  1 NA 0.0
#> [43,]  1 NA NA  NA
#> [44,]  1  1  1 1.0
#> [45,] NA  1  1 1.0
#> [46,]  0  1  1 1.0
#> [47,]  0  0 NA  NA
#> [48,]  0  0  0  NA
#> [49,] NA NA  1  NA
#> [50,]  0  1  1 1.0
#> [51,] NA  1  1 1.0
#> [52,]  1 NA NA 1.0
#> [53,] NA NA  1 1.0
#> [54,]  0  1  0 1.0
#> [55,]  0  0 NA  NA
#> [56,]  0  0  0 1.0
#> [57,]  1  1  1 1.0
#> [58,]  1  1 NA 1.0
#> [59,] NA  1  0 1.0
#> [60,] NA  1  1 1.0
#> [61,]  1  1 NA 1.0
#> [62,] NA NA  1 1.0
#> [63,]  0  0 NA  NA
#> [64,]  1 NA  1 1.0
#> [65,]  1 NA  1  NA
#> [66,] NA  1  0  NA
#> [67,]  1  1  1 1.0
#> [68,] NA  1  1  NA
#> [69,] NA NA NA 1.0
#> [70,] NA  1 NA 1.0
#> [71,]  1 NA  1 1.0
#> [72,] NA NA  0  NA
#> 
#> [[3]]
#>        C1  C2  C3 C4
#>  [1,]  NA 1.0  NA NA
#>  [2,]  NA 0.0  NA  0
#>  [3,] 0.0 1.0  NA NA
#>  [4,] 1.0  NA  NA NA
#>  [5,] 1.0  NA  NA NA
#>  [6,] 0.5 0.0  NA  0
#>  [7,] 1.0 0.0  NA  0
#>  [8,] 0.5  NA 1.0 NA
#>  [9,] 1.0  NA  NA NA
#> [10,] 1.0  NA 1.0  1
#> [11,] 1.0  NA  NA NA
#> [12,]  NA 1.0  NA NA
#> [13,] 0.0  NA  NA NA
#> [14,]  NA  NA 1.0  1
#> [15,] 1.0 1.0  NA  1
#> [16,]  NA 1.0 0.5  0
#> [17,] 1.0 1.0  NA  1
#> [18,] 1.0  NA 1.0 NA
#> [19,] 1.0 1.0 1.0 NA
#> [20,]  NA  NA 1.0 NA
#> [21,] 1.0 1.0 1.0  1
#> [22,] 1.0 1.0 1.0  1
#> [23,]  NA 0.0  NA  0
#> [24,]  NA  NA 0.0 NA
#> [25,]  NA  NA 0.0 NA
#> [26,]  NA  NA 1.0 NA
#> [27,] 0.0 0.0 0.0 NA
#> [28,] 1.0 1.0 1.0  1
#> [29,] 1.0  NA 1.0  1
#> [30,] 1.0  NA  NA NA
#> [31,] 1.0  NA 1.0 NA
#> [32,]  NA  NA 0.0 NA
#> [33,]  NA 0.0  NA  0
#> [34,] 1.0 0.0 0.0  0
#> [35,]  NA 1.0  NA NA
#> [36,] 1.0 1.0 1.0  1
#> [37,] 0.0  NA  NA NA
#> [38,]  NA 0.0  NA NA
#> [39,] 0.0  NA 0.0  0
#> [40,]  NA 1.0 1.0  1
#> [41,]  NA 1.0 1.0 NA
#> [42,]  NA 1.0  NA  1
#> [43,] 1.0  NA  NA NA
#> [44,]  NA 1.0  NA  0
#> [45,]  NA  NA  NA  1
#> [46,] 1.0 0.5 0.0  0
#> [47,]  NA 1.0  NA  1
#> [48,]  NA 1.0  NA  1
#> [49,]  NA  NA 1.0 NA
#> [50,]  NA  NA  NA  1
#> [51,]  NA  NA  NA  1
#> [52,] 0.0 0.0 0.0  0
#> [53,]  NA  NA 0.0 NA
#> [54,] 1.0  NA 0.0  0
#> [55,]  NA  NA  NA  1
#> [56,]  NA  NA 0.0  0
#> [57,] 0.0 0.0 0.0  0
#> [58,] 0.0 0.0 0.0  0
#> [59,] 0.0 0.0 0.0  0
#> [60,] 0.0 0.0 0.0  0
#> [61,]  NA  NA 1.0 NA
#> [62,]  NA 1.0  NA NA
#> [63,]  NA 1.0 1.0 NA
#> [64,] 1.0  NA  NA  1
#> [65,] 1.0  NA 1.0 NA
#> [66,] 1.0 1.0 1.0 NA
#> [67,]  NA 1.0 1.0  1
#> [68,] 0.0 1.0 1.0  0
#> [69,]  NA 1.0 1.0  1
#> [70,] 1.0 1.0 1.0  1
#> [71,]  NA 1.0  NA NA
#> 
#> [[4]]
#>        C1  C2 C3  C4
#>  [1,] 0.0 0.0 NA 0.0
#>  [2,] 1.0 1.0 NA  NA
#>  [3,] 1.0  NA NA 0.0
#>  [4,]  NA 1.0 NA  NA
#>  [5,]  NA  NA  1 0.5
#>  [6,] 1.0  NA NA 1.0
#>  [7,]  NA  NA  1  NA
#>  [8,]  NA 1.0 NA  NA
#>  [9,]  NA 1.0  0 1.0
#> [10,] 1.0 1.0  1 1.0
#> [11,] 1.0 1.0 NA 1.0
#> [12,] 0.0  NA  0  NA
#> [13,] 1.0 1.0  1  NA
#> [14,] 0.0  NA NA  NA
#> [15,] 1.0 1.0 NA 0.0
#> [16,] 1.0 1.0  1  NA
#> [17,]  NA 1.0 NA 1.0
#> [18,] 1.0 0.0  1 1.0
#> [19,] 1.0 1.0  1  NA
#> [20,] 1.0 1.0  1 1.0
#> [21,]  NA  NA  1 1.0
#> [22,] 1.0  NA NA  NA
#> [23,] 1.0 1.0 NA 1.0
#> [24,] 1.0 1.0  0 0.0
#> [25,] 1.0 1.0  0 0.0
#> [26,] 1.0 1.0  1 1.0
#> [27,] 0.0 1.0 NA 0.0
#> [28,]  NA 1.0  1 0.0
#> [29,] 1.0  NA  1 1.0
#> [30,] 1.0 0.0  0 0.0
#> [31,]  NA  NA  1 1.0
#> [32,] 1.0 0.0  1 0.0
#> [33,]  NA  NA  1 1.0
#> [34,]  NA  NA NA 1.0
#> [35,] 0.0  NA  1  NA
#> [36,]  NA 1.0  1  NA
#> [37,] 1.0 1.0  1  NA
#> [38,] 1.0 1.0 NA  NA
#> [39,] 1.0  NA NA  NA
#> [40,] 1.0  NA  1  NA
#> [41,] 0.0 1.0  1  NA
#> [42,] 1.0 1.0  1  NA
#> [43,]  NA  NA  0  NA
#> [44,] 1.0 1.0  1  NA
#> [45,]  NA 1.0  1  NA
#> [46,] 0.0 0.0  1 0.0
#> [47,]  NA 1.0  1 1.0
#> [48,] 1.0 1.0  1 1.0
#> [49,] 1.0  NA  1 1.0
#> [50,] 1.0  NA  1 1.0
#> [51,] 0.0  NA NA 1.0
#> [52,] 1.0 1.0 NA  NA
#> [53,]  NA  NA  1  NA
#> [54,]  NA  NA NA 0.0
#> [55,] 0.0  NA NA 1.0
#> [56,] 0.0 1.0 NA 1.0
#> [57,] 1.0 1.0 NA 1.0
#> [58,] 1.0 1.0  0  NA
#> [59,] 0.5 0.5  1  NA
#> [60,]  NA  NA  0 0.0
#> [61,]  NA 1.0 NA 0.0
#> [62,] 1.0 1.0 NA 1.0
#> [63,] 0.0  NA  0 0.0
#> [64,] 1.0 1.0  1 1.0
#> [65,]  NA  NA NA 0.0
#> [66,] 0.0 0.0  0 0.0
#> [67,] 0.0 0.0  0 0.0
#> [68,] 0.0 0.0  0 0.0
#> [69,] 0.0 0.0  0 0.0
#> [70,] 0.0 0.0  0 0.0
#> [71,] 0.0 0.0  0 0.0
#>